Update MPAS-A dycore to MPAS-Atmosphere v7.0
This commit updates the MPAS-A external to point to a branch off of the v7.0+
code (develop branch plus CAM-MPAS changes that have yet to be merged to
develop).

Also included in this commit are the addition of new namelist options in MPAS-A
v7 and minor updates to the MPAS-A dycore Makefile to handle new source files.